+++ brotherwillies/settings.py
@@ -60,6 +60,7 @@ TEMPLATES = [
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
                 'apps.accounts.context_processors.user_profile',
+                'apps.ops.context_processors.ops_status',
             ],
         },
     },